scraper/requirements.txt
# Lista bibliotek Pythona

requests
beautifulsoup4
python-dotenv
supabase
icalendar
tqdm